requirements.txt
streamlit
pandas
matplotlib
scikit-learn
google-generativeai
